rewritten session driver code to support concurrent requests ran all drivers through a 4 hour concurrency test (5 browser windows loading the same URL with random processing delays)

+++ fuel/core/classes/session/db.php
@@ -18,16 +18,24 @@ namespace Fuel;
 
 class Session_Db extends Session_Driver {
 
+	/*
+	 * @var	session database result object
+	 */
+	protected $record = null;
+
 	// --------------------------------------------------------------------
 
 	/**
-	 * driver initialisation
+	 * database driver initialisation
 	 *
 	 * @access	public
 	 * @return	void
 	 */
 	public function init()
 	{
+		// generic driver initialisation
+		parent::init();
+
 		// get the active database if needed
 		if ( ! isset($this->config['database']))
 		{
@@ -36,10 +44,88 @@ class Session_Db extends Session_Driver {
 		}
 
 		// check for required config values
+		$this->config['cookie_name'] = $this->validate_config('cookie_name', isset($this->config['cookie_name'])
+				? $this->config['cookie_name'] : 'fueldid');
 		$this->config['database'] = $this->validate_config('database', $this->config['database']);
 		$this->config['table'] = $this->validate_config('table', isset($this->config['table']) ? $this->config['table'] : null);
 		$this->config['gc_probability'] = $this->validate_config('gc_probability', isset($this->config['gc_probability'])
 				? $this->config['gc_probability'] : 5);
+	}
+
+	// --------------------------------------------------------------------
+
+	/**
+	 * create a new session
+	 *
+	 * @access	public
+	 * @return	void
+	 */
+	public function create()
+	{
+		// create a new session
+		$this->keys['session_id']	= $this->_new_session_id();
+		$this->keys['previous_id']	= $this->keys['session_id'];	// prevents errors if previous_id has a unique index
+		$this->keys['ip_address']	= Input::real_ip();
+		$this->keys['user_agent']	= Input::user_agent();
+		$this->keys['created'] 		= $this->time->get_timestamp();
+		$this->keys['updated'] 		= $this->keys['created'];
+		$this->keys['payload'] 		= '';
+
+		// create the session record
+		$result = DB::insert($this->config['table'], array_keys($this->keys))->values($this->keys)->execute($this->config['database']);
+
+		// and set the session cookie
+		$this->_set_cookie();
+	}
+
+	// --------------------------------------------------------------------
+
+	/**
+	 * read the session
+	 *
+	 * @access	public
+	 * @param	boolean, set to true if we want to force a new session to be created
+	 * @return	void
+	 */
+	public function read($force = false)
+	{
+		// get the session cookie
+		$cookie = $this->_get_cookie();
+
+		// if no session cookie was present, create it
+		if ($cookie === false or $force)
+		{
+			$this->create();
+		}
+
+		// read the session record
+		$this->record = DB::select()->where('session_id', '=', $this->keys['session_id'])->from($this->config['table'])->execute($this->config['database']);
+
+		// record found?
+		if ($this->record->count())
+		{
+			$payload = $this->_unserialize($this->record->get('payload'));
+		}
+		else
+		{
+			// try to find the session on previous id
+			$this->record = DB::select()->where('previous_id', '=', $this->keys['session_id'])->from($this->config['table'])->execute($this->config['database']);
+
+			// record found?
+			if ($this->record->count())
+			{
+				$payload = $this->_unserialize($this->record->get('payload'));
+			}
+			else
+			{
+				// cookie present, but session record missing. force creation of a new session
+				$this->read(true);
+				return;
+			}
+		}
+
+		if (isset($payload[0])) $this->data = $payload[0];
+		if (isset($payload[1])) $this->flash = $payload[1];
 	}
 
 	// --------------------------------------------------------------------
@@ -52,15 +138,36 @@ class Session_Db extends Session_Driver {
 	 */
 	public function write()
 	{
-		parent::write();
-
-		// do some garbage collection
-		srand(time());
-		if ((rand() % 100) < $this->config['gc_probability'])
+		// do we have something to write?
+		if ( ! empty($this->keys) and ! empty($this->record))
 		{
-			$time = Date::time();
-			$expired = $time->get_timestamp() - $this->config['expiration_time'];
-			$result = DB::delete($this->config['table'])->where('updated', '<', $expired)->execute($this->config['database']);
+			// rotate the session id if needed
+			$this->rotate(false);
+
+			// create the session record, and add the session payload
+			$session = $this->keys;
+			$session['payload'] = $this->_serialize(array($this->data, $this->flash));
+
+			// update the database
+			$result = DB::update($this->config['table'])->set($session)->where('session_id', '=', $this->record->get('session_id'))->execute($this->config['database']);
+
+			// update went well?
+			if ($result)
+			{
+				// then update the cookie
+				$this->_set_cookie();
+			}
+			else
+			{
+				Log::error('Session update failed, session record could not be found. Concurrency issue?');
+			}
+
+			// do some garbage collection
+			if (mt_rand(0,100) < $this->config['gc_probability'])
+			{
+				$expired = $this->time->get_timestamp() - $this->config['expiration_time'];
+				$result = DB::delete($this->config['table'])->where('updated', '<', $expired)->execute($this->config['database']);
+			}
 		}
 	}
 
@@ -74,11 +181,16 @@ class Session_Db extends Session_Driver {
 	 */
 	public function destroy()
 	{
-		// delete the session record
-		$result = DB::delete($this->config['table'])->where('session_id', '=', $this->keys['session_id'])->execute($this->config['database']);
+		// do we have something to destroy?
+		if ( ! empty($this->keys) and ! empty($this->record))
+		{
+			// delete the session record
+			$result = DB::delete($this->config['table'])->where('session_id', '=', $this->keys['session_id'])->execute($this->config['database']);
+		}
 
-		// delete the cookie, reset all session variables
-		parent::destroy();
+		// reset the stored session data
+		$this->record = null;
+		$this->keys = $this->flash = $this->data = array();
 	}
 
 	// --------------------------------------------------------------------
@@ -95,13 +207,20 @@ class Session_Db extends Session_Driver {
 	{
 		switch ($name)
 		{
+			case 'cookie_name':
+				if ( empty($value) OR ! is_string($value))
+				{
+					$value = 'fueldid';
+				}
+			break;
+
 			case 'database':
 				// do we have a database?
 				if ( empty($value) OR ! is_string($value))
 				{
 					throw new Exception('You have specify a database to use database backed sessions.');
 				}
-				break;
+			break;
 
 			case 'table':
 				// and a table name?
@@ -109,7 +228,7 @@ class Session_Db extends Session_Driver {
 				{
 					throw new Exception('You have specify a database table name to use database backed sessions.');
 				}
-				break;
+			break;
 
 			case 'gc_probability':
 				// do we have a path?
@@ -118,113 +237,15 @@ class Session_Db extends Session_Driver {
 					// default value: 5%
 					$value = 5;
 				}
-				break;
+			break;
 
 			default:
-				break;
+			break;
 		}
 
 		// return the validated value
 		return $value;
 	}
-
-	// --------------------------------------------------------------------
-
-	/**
-	 * Sets or creates the session cookie
-	 *
-	 * the db driver stores data and flash in a database table
-	 *
-	 * @access	private
-	 * @return  void
-	 */
-	protected function _set_cookie($session_id = NULL)
-	{
-		// create the session cookie
-		parent::_set_cookie($session_id);
-
-		// get a time object
-		$time = Date::time();
-
-		// create the initial session record
-		$session = array(
-			'session_id' => $this->keys['session_id'],
-			'previous_id' => ( ! is_null($session_id) && $session_id != $this->keys['session_id']) ? $session_id : $this->keys['session_id'],
-			'ip_address' => Input::real_ip(),
-			'user_agent' => Input::user_agent(),
-			'updated' => $time->get_timestamp(),
-			'payload' => $this->_serialize(array($this->data, $this->flash))
-		);
-
-		if (is_null($session_id))
-		{
-			// set the create timestamp
-			$session['created'] = $time->get_timestamp();
-
-			// insert it into the database
-			$result = DB::insert($this->config['table'], array_keys($session))->values($session)->execute($this->config['database']);
-		}
-		else
-		{
-			// was the session id rotated?
-			if ( ! is_null($session_id) && $session_id != $this->keys['session_id'])
-			{
-				// if so, save the previous id
-				$session['previous_id'] = $session_id;
-			}
-			// update the database
-			$result = DB::update($this->config['table'])->set($session)->where('session_id', '=', $session_id)->execute($this->config['database']);
-		}
-	}
-
-	// --------------------------------------------------------------------
-
-	/**
-	 * Gets the session cookie
-	 *
-	 * the db driver stores data and flash in a database table
-	 *
-	 * @access	private
-	 * @return  boolean, true if found, false if not
-	 */
-	protected function _get_cookie()
-	{
-		// fetch the session cookie
-		parent::_get_cookie();
-
-		// read the session record
-		if (isset($this->keys['session_id']))
-		{
-			$result = DB::select()->where('session_id', '=', $this->keys['session_id'])->from($this->config['table'])->execute($this->config['database']);
-
-			// record found?
-			if (count($result))
-			{
-				$payload = $this->_unserialize($result->get('payload'));
-			}
-			else
-			{
-				// try to find the session on previous id
-				$result = DB::select()->where('previous_id', '=', $this->keys['session_id'])->from($this->config['table'])->execute($this->config['database']);
-
-				// record found?
-				if (count($result))
-				{
-					$payload = $this->_unserialize($result->get('payload'));
-				}
-				else
-				{
-					$this->keys = NULL;
-				}
-			}
-		}
-
-		if (isset($payload[0])) $this->data = $payload[0];
-		if (isset($payload[1])) $this->flash = $payload[1];
-
-		return ! empty($this->keys);
-	}
-
 }
 
-/* End of file driver.php */
+/* End of file db.php */
